Ingestion and schema inference updates * Refactor ingestion to use Confluent Kafka Client * Add add_fields_from_dataframe method instead of update_from_dataset * Refactor loaders module
